refactor: Session 헬퍼·로깅 설정 모듈 core 이주 (#62)
- core/api/session.py 신설: 공유 세션·스레드로컬 get_thread_session·쿠키
  저장 차단 정책을 content/network.py에서 그대로 이주 (동작 무변경)
- core/utils/logging_setup.py 신설: 콘솔+회전 파일 핸들러·urllib3 억제·
  중복 초기화 가드 이주. 로그 위치(CONFIG_DIR/logs)는 app 영역 지식이라
  core는 log_dir을 인자로 받고 config/log_setup.py 진입점이 주입
- content/network.py에 re-export 유지, download/*.py는 core 모듈 직접 사용
- 테스트를 tests/unit/core/{test_session,test_logging_setup}.py로 이주,
  기존 위치에는 진입점(경로 주입) 검증만 남김
